% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asi-package.R
\docType{package}
\name{asi-package}
\alias{asi}
\alias{asi-package}
\title{asi: Annual Survey of Industries (ASI)}
\description{
Microdata from the Annual Survey of Industries (ASI) conducted by the National Sample Survey Office (NSSO), Ministry of Statistics and Programme Implementation, Government of India. Tidied and organized using DuckDB for efficient querying.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rahulsh97/asi}
  \item Report bugs at \url{https://github.com/rahulsh97/asi/issues}
}

}
\author{
\strong{Maintainer}: Rahul Shukla \email{rahulshuklakrc@gmail.com}

}
\keyword{internal}
