% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.r
\name{asi_download}
\alias{asi_download}
\title{Download the Annual Survey of Industries database from GitHub (2015-16 to 2022-23)}
\usage{
asi_download(ver = NULL)
}
\arguments{
\item{ver}{Which version to download. By default it is the latest version available on GitHub. You can see all
versions at \url{https://github.com/rahulsh97/asi/releases}.}
}
\description{
This function downloads the database from GitHub and creates a local database (DuckDB) with the data. These
datasets cannot be included in the package due to their size and CRAN file size limits.
}
\examples{
\dontrun{ asi_download() }
}
