% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{asi_file_path}
\alias{asi_file_path}
\title{Local ASI database file path}
\usage{
asi_file_path(dir = asi_path())
}
\arguments{
\item{dir}{Path to the ASI directory on disk. By default this is #' "asi" inside the user's R data directory, or the
directory specified by the \code{ASI_DIR} environment variable if set.}
}
\description{
Returns the path to the local ASI database directory.
}
\examples{
asi_file_path()
}
